services/vanna/requirements.txt
vanna==0.7.7
groq==0.15.0
psycopg2-binary==2.9.9
fastapi==0.115.6
uvicorn[standard]==0.34.2
python-dotenv==1.0.1
pandas==2.2.3
